Split stdlib path resolution
